% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplineRDA.R
\name{splineGradient}
\alias{splineGradient}
\title{Does post-processing of canonical analysis object from vegan to extract spline gradients}
\usage{
splineGradient(object, Bdf, Gnames = NULL, predB = NULL, plot = T, ...)
}
\arguments{
\item{object}{The fitted rda/cca or capscale object}

\item{Bdf}{Integer vector giving the degrees of freedom (number of columns) of the bases used.}

\item{Gnames}{Gradient names corresponding to the bases.}

\item{predB}{List containing the new set of bases for prediction.}

\item{plot}{Draw plot of the gradient. predB must be supplied.}
}
\value{
Gradt Fitted gradient values

Env.scores Correlation scores for spline and linear covariates, for adding to a biplot

pred.Gradt List containing the fitted gradient values for predB
}
\description{
Does post-processing of canonical analysis object from vegan to extract spline gradients
}
\examples{
load("R/PitTraps.RData")
Locn=PitTraps$Locn
distance=PitTraps$distance
Counts=PitTraps[,-c(1,2)]
B=bSpline(distance,degree=2,df=5)   #Spline basis for distance
Bpred=predict(B,newx=min(distance):max(distance)) #For prediction
CAP.fit=capscale((Counts>0)~B+Locn,dist="bray") #Fit RDA on PCOs
Gradients=splineGradient(CAP.fit,Bdf=5,predB=list(Bpred),type="l")

}
